Modernize frontend AST handling and test generation (#810)
## Summary

This branch modernizes a broad compiler cleanup pass across the Clang
frontend, Sage/OpenMP AST construction, unparsing, analysis support, and
functional test registration.

The main behavioral fix is on the frontend side for OpenMP/OpenACC
parsing and AST construction: explicit-end target regions are now
represented structurally in the AST, commented directive ordering is
preserved by AST/source ownership, and the unparser no longer relies on
text hacks to recover those constructs.

## What Changed

- Tightened Clang frontend lowering in `clang-frontend-decl.cpp`,
`clang-frontend-stmt.cpp`, `clang-frontend-type.cpp`, and related Sage
builder/support code so newer C/C++ forms are carried into Sage with
more accurate structure and source location data.
- Reworked OpenMP AST construction in
`src/frontend/SageIII/ompAstConstruction.cpp` to:
  - pair C/C++ explicit-end target directives in the frontend,
  - build real body blocks for those regions,
  - preserve the matching end pragma as syntax in the AST,
- relocate scope-owned preprocessing info/comments by source order
instead of string-based recovery.
- Updated unparser behavior in `src/backend/unparser/unparser.C`,
`src/backend/unparser/languageIndependenceSupport/unparseLanguageIndependentConstructs.C`,
and the C++ unparser files so OpenMP/OpenACC output is driven by AST
state rather than attribute/string fallbacks.
- Cleaned up preprocessing/comment attachment and related AST utilities
in `attachPreprocessingInfo.C`, `sage_support.C`, `sageInterface.C`,
call-graph code, and lowering helpers that depended on the older
behavior.
- Simplified functional test registration CMake by removing checked-in
generated `*_Testcodes.cmake` fragments and consolidating the source
lists directly in current `CMakeLists.txt` files.
- Added `cmake/check_no_autotools.cmake` and related CMake updates for
the modernized test/build flow.
- Regenerated affected move-declaration and SageInterface reference
outputs from the compiler-side fixes instead of masking regressions in
the tests.
- Removed generated test artifacts from the source tree so they no
longer remain as untracked pollution after test runs.

## Why

Two concrete OpenMP regressions were fixed as part of this pass:

- `OMPTEST_output_diff_variables_c`
- `OMPTEST_output_diff_jacobi_ompacc_multiGPU_c`

Both failures came from incorrect frontend ownership/placement of
commented pragmas. The fix is now structural: the frontend places the
directives and associated comments on the correct AST boundaries early,
and the unparser consumes that AST directly.

That same frontend-first approach also fixed the earlier explicit-end
target region regression (`axpy_ompacc_parseonly`) without synthesizing
end-directive text in the unparser.

## Validation

- `cmake --build build -j"$(nproc)"`
- `ctest --test-dir build --output-on-failure -R
'^OMPTEST_output_diff_variables_c$|^OMPTEST_output_diff_jacobi_ompacc_multiGPU_c$|^OMPTEST_output_variables_c$|^OMPTEST_output_jacobi_ompacc_multiGPU_c$|^OMPTEST_variables.c$|^OMPACCTEST_jacobi-ompacc-multiGPU.c$'`
- `ctest --test-dir build -R
'rex|astInterface|testQuery|fortran|f90|f03|gfortran|OMPTEST_|OMPACCTEST_|OMPFORTRAN_|OMPANALYZE_|OMPVV_5_0_|OMPVV_4_5_|OMPVV_5_1_|OMPVV_5_2_|OMPVV_6_0_|omp_lowering_|OMPLOWERING_CPU_|OMPLOWERING_RODINIA_'
--output-on-failure -j"$(nproc)"`
  - Passed locally in full: `7151/7151` tests
- `pre-commit` hook ran `git clang-format --staged`
- `pre-push` hook rebuilt and reran the same CI ctest selection
successfully before push

+++ tests/nonsmoke/functional/moveDeclarationTool/referenceResults/withmerge/rose_v2_inputmoveDeclarationToInnermostScope_9.C
@@ -2,8 +2,8 @@
 
 void fallbackSimpleSort(int *fmap,int *eclass,int lo,int hi)
 {
-  if (lo == hi) 
-    return ;
+  if (lo == hi)
+    return;
   if (hi - lo > 3) {
     int i;
     unsigned int ec_tmp;
@@ -19,7 +19,7 @@ void fallbackSimpleSort(int *fmap,int *eclass,int lo,int hi)
   for (int i = hi - 1; i >= lo; i--) {
     int j;
     int tmp = fmap[i];
-    unsigned int ec_tmp = (unsigned int )eclass[tmp];
+    unsigned int ec_tmp = eclass[tmp];
     for (j = i + 1; j <= hi && ec_tmp > eclass[fmap[j]]; j++) 
       fmap[j - 1] = fmap[j];
     fmap[j - 1] = tmp;
